Create boxmodel_solver_debug.F90
boxmodel_solver_debug is a version working on element nodes and not on the element index as it was the case for the SSA version.

Version should be called "boxmodel3D_solver" or similar when finished.

Model gives similar melt to the "element" version but Salinity seems particularly high and Temperature particularly low in the the k>1 boxes